separate voronoi to another module
