+++ bet/postProcess/__init__.py
@@ -10,4 +10,4 @@ This subpackage contains
 * :class:`~bet.postProcess.postTools` has tools for postprocessing
 
 """
-__all__ = ['plotP', 'plotDomains', 'postTools']
+__all__ = ['plotP', 'plotDomains', 'postTools', 'plotVoronoi']
