+++ test/test_postProcess/__init__.py
@@ -1,1 +1,2 @@
-__all__ = ["test_plotDomains", "test_postTools", "test_plotP"]
+__all__ = ["test_plotDomains", "test_postTools", "test_plotP",
+           "test_plotVoronoi"]
